Fix timer firing even after signal has been received
There should not be any impact from the timer firing,
but for a cleaner event history, we should short-circuit,
and clear any pending futures.

The workflow (parent / caller) should also define a child context
that can be cancelled should the timer fire / timeout (though,
this is not necessary).

We are using `ReceiveAsync` after the `Select` to handle potential
race conditions in case the worker dies, and both branches passes.