ReceiveWithTimeout: check if the context is done
Prior to this commit, `ReceiveWithTimeout` would return `true`
when a workflow is cancelled as we did not account for a "done"
channel.

This commit checks if the context is "done", and include the
result as part of the new return value (a struct) so callers
can better disambiguate between actual timeouts, and the
context simply being cancelled.